+++ packages/upup/src/shared/i18n/index.ts
@@ -1,0 +1,3 @@
+export { en_US } from './en_US'
+export type { Translations } from './types'
+export { mergeTranslations, plural, t } from './utils'